requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2023.12.2
future==0.18.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
